+++ src/main/java/kitchenpos/eatinorder/application/EatInOrderService.java
@@ -8,8 +8,8 @@ import kitchenpos.eatinorder.domain.OrderTable;
 import kitchenpos.eatinorder.domain.OrderTableRepository;
 import kitchenpos.eatinorder.domain.OrderType;
 import kitchenpos.deliveryorder.infra.KitchenridersClient;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.menu.domain.Menu;
+import kitchenpos.menu.menu.domain.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
