+++ src/main/java/kitchenpos/menu/menu/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.infra;
+package kitchenpos.menu.menu.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
