+++ src/main/java/kitchenpos/menu/menugroup/domain/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.menugroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
