+++ src/test/java/kitchenpos/application/FakeMenuPurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.menu.infra.PurgomalumClient;
+import kitchenpos.menu.menu.infra.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
